Rewrite facetline to follow the structure introduced in the last merge
This is to avoid duplicate code and logic for facet browser, facetline and campaigns
